requirements.txt
fastapi
uvicorn
playwright
beautifulsoup4
chromadb
openai
python-dotenv
tiktoken
